refactor: step1, 2의 App.js, controller 파일 디렉터리 분리
