+++ src/step1-js/LottoGameController.js
@@ -1,9 +1,9 @@
-import InputView from './view/InputView.js';
-import OutputView from './view/OutputView.js';
-import Validation from './domain/Vaildation.js';
-import LottoGame from './domain/LottoGame.js';
-import Console from './utils/Console.js';
-import { LOTTO_CONDITION, RESTART_COMMAND } from './constants/condition.js';
+import InputView from '../view/InputView.js';
+import OutputView from '../view/OutputView.js';
+import Validation from '../domain/Vaildation.js';
+import LottoGame from '../domain/LottoGame.js';
+import Console from '../utils/Console.js';
+import { LOTTO_CONDITION, RESTART_COMMAND } from '../constants/condition.js';
 
 export default class LottoGameController {
   #lottoGame;
