+++ src/step2-js/App.js
@@ -1,0 +1,1 @@
+export default class App {}
